+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add Landing Page
 - Add User Account's registration, login and logout pages.
 - Add Volume's download link to `patterns/template-taxonomy-volumes.php`
 ### Changed
 - `functions.php` has been refactored by separating grouped functionalities
 in separate files placed inside the `./inc/` directory.
+- Page title is more SEO friendly
 - Removed autofocus on search field when the menu is opened in `script.js`
 
 ## [0.4.0] 2025-07-13